refactor: 使用 Koin 重构数据层
