+++ app/src/main/kotlin/cn/super12138/todo/logic/datastore/DataStoreManager.kt
@@ -1,7 +1,9 @@
 package cn.super12138.todo.logic.datastore
 
 import android.content.Context
+import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.SharedPreferencesMigration
+import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.floatPreferencesKey
@@ -14,21 +16,7 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import kotlinx.serialization.json.Json
 
-object DataStoreManager {
-    private val Context.dataStore by preferencesDataStore(
-        name = Constants.SP_NAME,
-        produceMigrations = { context ->
-            listOf(
-                SharedPreferencesMigration(
-                    context = context,
-                    sharedPreferencesName = Constants.SP_NAME,
-                )
-            )
-        }
-    )
-
-    val dataStore = VerveDoApp.context.dataStore
-
+class DataStoreManager(val dataStore: DataStore<Preferences>) {
     // Keys
     // 外观与个性化
     private val DYNAMIC_COLOR = booleanPreferencesKey(Constants.PREF_DYNAMIC_COLOR)
