+++ app/src/main/kotlin/cn/super12138/todo/logic/model/ScreenMode.kt
@@ -1,0 +1,7 @@
+package cn.super12138.todo.logic.model
+
+enum class ScreenMode {
+    Normal, // 正常模式
+    Selection, // 选择模式
+    Search, // 搜索模式
+}
