+++ app/src/main/kotlin/cn/super12138/todo/ui/pages/tasks/TaskStateHolder.kt
@@ -1,0 +1,4 @@
+package cn.super12138.todo.ui.pages.tasks
+
+class TaskStateHolder {
+}
