+++ app/src/main/kotlin/cn/super12138/todo/ui/viewmodels/EditorViewModel.kt
@@ -1,0 +1,7 @@
+package cn.super12138.todo.ui.viewmodels
+
+import androidx.lifecycle.ViewModel
+import cn.super12138.todo.logic.datastore.DataStoreManager
+
+class EditorViewModel(private val dataStoreManager: DataStoreManager) : ViewModel() {
+}
