+++ gradle/libs.versions.toml
@@ -10,7 +10,6 @@ composeBom = "2026.03.01"
 liveData = "1.10.6"
 room = "2.8.4"
 material3 = "1.5.0-alpha17"
-adaptive = "1.2.0"
 nav3Core = "1.1.0"
 # AboutLibraries
 aboutLibsRelease = "14.0.1"
@@ -22,10 +21,12 @@ m3color = "2026.1"
 # Konfetti
 konfetti = "2.0.5"
 # Lazy Column Scrollbar
-lazyColumnScrollbar = "2.2.0"
+# lazyColumnScrollbar = "2.2.0"
 # Kotlin
 kotlinCoroutines = "1.10.2"
 kotlinSerialization = "1.11.0"
+# Koin
+koin-bom = "4.2.1"
 # Test
 junit = "4.13.2"
 junitVersion = "1.3.0"
@@ -82,6 +83,12 @@ nl-dionsegijn-konfetti-compose = { group = "nl.dionsegijn", name = "konfetti-com
 kotlinx-coroutines-core = { group = "org.jetbrains.kotlinx", name = "kotlinx-coroutines-core", version.ref = "kotlinCoroutines" }
 kotlinx-coroutines-android = { group = "org.jetbrains.kotlinx", name = "kotlinx-coroutines-android", version.ref = "kotlinCoroutines" }
 kotlinx-serialization-json = { group = "org.jetbrains.kotlinx", name = "kotlinx-serialization-json", version.ref = "kotlinSerialization" }
+
+# Koin
+koin-bom = { group = "io.insert-koin", name = "koin-bom", version.ref = "koin-bom" }
+koin-android = { group = "io.insert-koin", name = "koin-android" }
+koin-compose = { group = "io.insert-koin", name = "koin-compose" }
+koin-compose-viewmodel = { group = "io.insert-koin", name = "koin-compose-viewmodel" }
 
 # Test
 junit = { group = "junit", name = "junit", version.ref = "junit" }
